requirements.txt
Flask
bcrypt
Pillow
pdf2docx
docx2pdf
python-magic
uuid
